+++ andropilot/src/main/scala/com/geeksville/gmaps/SmartMarker.scala
@@ -106,37 +106,3 @@ abstract class SmartMarker extends AndroidLogger {
   def showInfoWindow() = gmarker.foreach(_.showInfoWindow())
 }
 
-trait LineFactory {
-  var polyline: Option[Polyline] = None
-
-  def lineOptions: PolylineOptions
-
-  /**
-   * Move any segments as we are dragged (if we care)
-   */
-  def handleMarkerDrag(sm: SmartMarker) {}
-}
-
-/**
- * A line between two SmartMarkers
- */
-case class Segment(endpoints: (SmartMarker, SmartMarker), color: Int) extends LineFactory {
-  final def lineOptions = (new PolylineOptions).color(color).add(endpoints._1.latLng).add(endpoints._2.latLng)
-
-  /**
-   * Move any segments as we are dragged (if we care)
-   */
-  override def handleMarkerDrag(sm: SmartMarker) {
-    if (endpoints._1 == sm || endpoints._2 == sm) {
-      val points = List(endpoints._1, endpoints._2).map(_.latLng)
-      polyline.foreach(_.setPoints(points.asJava))
-    }
-  }
-}
-
-/**
- * Just a series of poly points
- */
-case class PolylineFactory(points: Iterable[LatLng], color: Int) extends LineFactory {
-  final def lineOptions = (new PolylineOptions).color(color).addAll(points.asJava)
-}