+++ common/src/main/scala/com/geeksville/flight/ParametersReadOnlyModel.scala
@@ -101,6 +101,9 @@ trait ParametersReadOnlyModel extends MavlinkConstants {
     }
 
     def getInt = raw.map { v => v.param_value.toInt }
+    def getFloat = raw.map { v => v.param_value }
+
+    def getBoolean = getInt.map(_ != 0)
 
     /**
      * @return a nice human readable version of this value (decoding based on documentation if possible)